address remarks from code review
