+++ assume/markets/__init__.py
@@ -1,1 +1,2 @@
+from assume.markets.base_market import MarketRole
 from assume.markets.clearing_algorithms import clearing_mechanisms
